src/main/java/com/api/prog4td1p1/ClassSingleton.java
package com.api.prog4td1p1;

import lombok.AccessLevel;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor(access = AccessLevel.PRIVATE)
public final class ClassSingleton {

    private static ClassSingleton INSTANCE;

    private String description = "I'm a singleton, the one and only";
    public static ClassSingleton getInstance() {
        if(INSTANCE == null) {
            INSTANCE = new ClassSingleton();
        }

        return INSTANCE;
    }
}